requirements.txt
asgiref==3.7.2
attrs==23.2.0
autopep8==2.0.4
certifi==2024.2.2
coverage==7.3.2
Django==4.2.6
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.1.3
Faker==19.10.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
mypy==1.6.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
parameterized==0.9.0
Pillow==10.1.0
pluggy==1.3.0
pycodestyle==2.11.1
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.2
pytest-django==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
